requirements.txt
torch==1.9.0
matplotlib==3.3.4
pandas==1.3.0
opencv_python==4.5.2.54
numpy==1.21.0
streamlit==0.84.0
Pillow==8.3.0
streamlit_webrtc==0.23.6
